requirements.txt
celery
dj-database-url
djangorestframework
Django
gunicorn
psycopg2-binary
whitenoise